+++ proxy/constants/constants.go
@@ -1,0 +1,8 @@
+package constants
+
+// Defines the maximum size for requests
+const (
+	MB = 1048576
+	MaxRequestSize = 1 * MB
+	MaxFileUploadSize = 16 * MB
+)
